+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ## 1.3.4 - November 28, 2021
 
+- Emphasized errors with read background and black text.
+- Added Sharedkey durign reset. Fixes VPN connection
 - Added dns and dhcp option in config; auto builds values for azure vm
 - Added vyatta cmd function; provides a in prompt cmd's remotely if needed
 - BETA: working support for latest vyos image